refactor(doctor): extract allowlist policy repair
